--- Okane.Client/src/features/financeRecords/constants/searchFilters.ts
@@ -1,4 +1,0 @@
-// Internal
-import type { FinanceRecordsSearchFilters } from '@features/financeRecords/types/searchFilters'
-
-export const DEFAULT_FINANCE_RECORD_SEARCH_FILTERS: FinanceRecordsSearchFilters = {}
