+++ Okane.Client/src/features/financeRecords/constants/searchFinanceRecords.ts
@@ -1,0 +1,4 @@
+// Internal
+import type { FinanceRecordsSearchFilters } from '@features/financeRecords/types/searchFinanceRecords'
+
+export const DEFAULT_FINANCE_RECORD_SEARCH_FILTERS: FinanceRecordsSearchFilters = {}
